+++ sparse_strips/vello_common/src/blurred_rounded_rect.rs
