+++ sparse_strips/vello_common/src/lib.rs
@@ -15,7 +15,7 @@ only break in edge cases, and some of them are also only related to conversions 
 
 extern crate alloc;
 
-pub mod blurred_rect;
+pub mod blurred_rounded_rect;
 pub mod coarse;
 pub mod encode;
 pub mod flatten;
